% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmsterdamChess-data.R
\docType{data}
\name{AmsterdamChess}
\alias{AmsterdamChess}
\title{Amsterdam Chess Test (ACT) data}
\format{
A dataframe with 259 rows and 81 variables.
}
\usage{
data(AmsterdamChess)
}
\description{
Responses and response time data from the Amsterdam Chess Test (ACT).
}
\details{
\itemize{
  \item ELO: Standardized ELO rating (numeric)
  \item A1-A40: item correct score (1 or 0) for scored items 1 – 40 (numeric)
  \item AR1-AR40: response time (seconds) for scored items 1 – 40 (numeric)
}
}
\references{
van der Maas, H. L., & Wagenmakers, E. J. (2005). A psychometric analysis of chess expertise. The American journal of psychology, 118(1), 29-60.
(\href{https://pubmed.ncbi.nlm.nih.gov/15822609}{PubMed})
}
\keyword{datasets}
