% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CredentialForm1-data.R
\docType{data}
\name{CredentialForm1}
\alias{CredentialForm1}
\title{Credential Form data}
\format{
A dataframe with 1636 rows and 610 variables.
}
\usage{
data(CredentialForm1)
}
\description{
Responses and response time data from the credential data set of Cizek and Wollack (2016).
}
\references{
Cizek GJ, Wollack JA (eds.) (2016). Handbook of Quantitative Methods for Detecting Cheating on Tests. Routledge.
(\href{https://www.taylorfrancis.com/books/9781315743097}{Taylor&Francis})
}
\keyword{datasets}
