% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach start message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{defunct}

\item{pkgname}{defunct}
}
\value{
invisible()
}
\description{
.onAttach start message
}
