% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LNIRT_Q.R
\name{LNIRTQ}
\alias{LNIRTQ}
\title{Log-normal response time IRT modelling with variable person speed (intercept, trend, quadratic)}
\usage{
LNIRTQ(Y, RT, X, XG = 1000, burnin = 10)
}
\arguments{
\item{Y}{a Person-x-Item matrix of responses.}

\item{RT}{a Person-x-Item matrix of log-response times (time spent on solving an item).}

\item{X}{explanatory (time) variables for random person speed (default: (1:N.items - 1)/N.items).}

\item{XG}{the number of MCMC iterations to perform (default: 1000).}

\item{burnin}{the percentage of MCMC iterations to discard as burn-in period (default: 10).}
}
\value{
an object of class LNIRTQ.
}
\description{
Log-normal response time IRT modelling with variable person speed (intercept, trend, quadratic)
}
