% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLNIRT.R
\name{simLNIRT}
\alias{simLNIRT}
\title{Simulate data for log-normal response time IRT modelling}
\usage{
simLNIRT(N, K, rho, td, WL)
}
\arguments{
\item{N}{the number of persons.}

\item{K}{the number of items.}

\item{rho}{the correlation between the person ability and person speed parameter.}

\item{td}{set time-discrimination to one (default: false).}

\item{WL}{define the time-discrimination parameter as measurement error variance parameter (default: false).}
}
\value{
an object of class simLNIRT.
}
\description{
Simulate data for log-normal response time IRT modelling
}

