% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLNIRT_Q.R
\name{simLNIRTQ}
\alias{simLNIRTQ}
\title{Simulate data for log-normal response time IRT modelling with variable person speed (intercept, trend, quadratic)}
\usage{
simLNIRTQ(N, K, ...)
}
\arguments{
\item{N}{the number of persons.}

\item{K}{the number of items.}

\item{...}{optional arguments.}
}
\value{
an object of class simLNIRTQ.
}
\description{
Simulate data for log-normal response time IRT modelling with variable person speed (intercept, trend, quadratic)
}
