% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMMARY_LNIRT_Q.R
\name{summaryIRTQ}
\alias{summaryIRTQ}
\title{Summary Function for LNIRTQ}
\usage{
summaryIRTQ(out, data)
}
\arguments{
\item{out}{a LNIRTQ object (the fitted model)}

\item{data}{a simLNIRTQ object (the simulated data, optional)}
}
\description{
Summary Function for LNIRTQ
}
