% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp_summary.R
\name{summaryRT}
\alias{summaryRT}
\title{summaryRT}
\usage{
summaryRT(out, data)
}
\arguments{
\item{out}{estimation result}

\item{data}{simulated data}
}
\description{
summary function for LNRT, and LNGAMMA
}

